Move span and other related classes to right package
