+++ opentelemetry-kotlin-compat/api/opentelemetry-kotlin-compat.api
@@ -10,6 +10,10 @@ public final class io/embrace/opentelemetry/kotlin/context/ContextExtKt {
 	public static final fun toOtelJavaContext (Lio/embrace/opentelemetry/kotlin/context/Context;)Lio/opentelemetry/context/Context;
 }
 
+public final class io/embrace/opentelemetry/kotlin/context/ContextKeyExtKt {
+	public static final fun toOtelJavaContextKey (Lio/embrace/opentelemetry/kotlin/context/ContextKey;)Lio/opentelemetry/context/ContextKey;
+}
+
 public final class io/embrace/opentelemetry/kotlin/context/OtelJavaContextExtKt {
 	public static final fun toOtelKotlinContext (Lio/opentelemetry/context/Context;)Lio/embrace/opentelemetry/kotlin/context/Context;
 }
@@ -26,16 +30,16 @@ public final class io/embrace/opentelemetry/kotlin/logging/export/OtelJavaLogRec
 	public static final fun toOtelKotlinLogRecordExporter (Lio/opentelemetry/sdk/logs/export/LogRecordExporter;)Lio/embrace/opentelemetry/kotlin/logging/export/LogRecordExporter;
 }
 
-public final class io/embrace/opentelemetry/kotlin/tracing/SpanContextStorageExtKt {
-	public static final fun storeInContext (Lio/embrace/opentelemetry/kotlin/tracing/model/Span;Lio/embrace/opentelemetry/kotlin/context/Context;)Lio/embrace/opentelemetry/kotlin/context/Context;
-}
-
 public final class io/embrace/opentelemetry/kotlin/tracing/export/OtelJavaSpanExporterExtKt {
 	public static final fun toOtelKotlinSpanExporter (Lio/opentelemetry/sdk/trace/export/SpanExporter;)Lio/embrace/opentelemetry/kotlin/tracing/export/SpanExporter;
 }
 
-public final class io/embrace/opentelemetry/kotlin/tracing/ext/ContextKeyExtKt {
-	public static final fun toOtelJavaContextKey (Lio/embrace/opentelemetry/kotlin/context/ContextKey;)Lio/opentelemetry/context/ContextKey;
+public final class io/embrace/opentelemetry/kotlin/tracing/ext/EventDataExtKt {
+	public static final fun toOtelJavaEventData (Lio/embrace/opentelemetry/kotlin/tracing/data/EventData;)Lio/opentelemetry/sdk/trace/data/EventData;
+}
+
+public final class io/embrace/opentelemetry/kotlin/tracing/ext/LinkDataExtKt {
+	public static final fun toOtelJavaLinkData (Lio/embrace/opentelemetry/kotlin/tracing/data/LinkData;)Lio/opentelemetry/sdk/trace/data/LinkData;
 }
 
 public final class io/embrace/opentelemetry/kotlin/tracing/ext/OtelJavaSpanContextExtKt {
@@ -44,5 +48,9 @@ public final class io/embrace/opentelemetry/kotlin/tracing/ext/OtelJavaSpanConte
 
 public final class io/embrace/opentelemetry/kotlin/tracing/ext/SpanContextExtKt {
 	public static final fun toOtelJavaSpanContext (Lio/embrace/opentelemetry/kotlin/tracing/model/SpanContext;)Lio/opentelemetry/api/trace/SpanContext;
+}
+
+public final class io/embrace/opentelemetry/kotlin/tracing/ext/SpanContextStorageExtKt {
+	public static final fun storeInContext (Lio/embrace/opentelemetry/kotlin/tracing/model/Span;Lio/embrace/opentelemetry/kotlin/context/Context;)Lio/embrace/opentelemetry/kotlin/context/Context;
 }
